put all the package manager commands in its own "namespace" (or command). (e.g. “$ cocos pkg install” instead of “$ cocos install”). complete package install
