+++ bin/cocos2d.ini
@@ -18,11 +18,8 @@ project_run.CCPluginRun
 project_deploy.CCPluginDeploy
 plugin_jscompile.CCPluginJSCompile
 plugin_luacompile.CCPluginLuaCompile
-package_search.CCPluginPackageSearch
-package_info.CCPluginPackageInfo
-package_install.CCPluginPackageInstall
-package_list.CCPluginPackageList
-framework_add.CCPluginFrameworkAdd
+package.CCPluginPackage
+framework.CCPluginFramework
 #plugin_version.CCPluginVersion
 #plugin_install.CCPluginInstall
 #plugin_update.CCPluginUpdate
