+++ plugins/framework/__init__.py
@@ -1,0 +1,2 @@
+
+from plugin_framework import CCPluginFramework
