--- plugins/framework_add/__init__.py
@@ -1,2 +1,0 @@
-
-from framework_add import CCPluginFrameworkAdd
