+++ plugins/package/__init__.py
@@ -1,0 +1,2 @@
+
+from plugin_package import CCPluginPackage
