--- plugins/package_info/__init__.py
@@ -1,2 +1,0 @@
-
-from package_info import CCPluginPackageInfo
