--- plugins/package_install/__init__.py
@@ -1,2 +1,0 @@
-
-from package_install import CCPluginPackageInstall
