--- plugins/package_list/__init__.py
@@ -1,2 +1,0 @@
-
-from package_list import CCPluginPackageList
