--- plugins/package_search/__init__.py
@@ -1,2 +1,0 @@
-
-from package_search import CCPluginPackageSearch
